+++ types/browser/inputValueFromControl.d.ts
@@ -1,0 +1,1 @@
+export { inputValueFromControl, inputValueFromControl as default } from "../browser.js";