+++ types/browser/matchesMediaQuery.d.ts
@@ -1,0 +1,1 @@
+export { matchesMediaQuery, matchesMediaQuery as default } from "../browser.js";