+++ types/browser/prefersColorScheme.d.ts
@@ -1,0 +1,1 @@
+export { prefersColorScheme, prefersColorScheme as default } from "../browser.js";