+++ types/browser/readJsonStorage.d.ts
@@ -1,0 +1,1 @@
+export { readJsonStorage, readJsonStorage as default } from "../browser.js";