+++ types/browser/writeJsonStorage.d.ts
@@ -1,0 +1,1 @@
+export { writeJsonStorage, writeJsonStorage as default } from "../browser.js";